#130 [FEAT] 하나의 역할만을 하는 소셜 프로필 카드 컴포넌트로 분리
