+++ components/social/profile/SocialProfileCard/SocialProfileCardView.tsx
@@ -1,9 +1,5 @@
 import styled from '@emotion/styled';
-import { useRouter } from 'next/router';
 
-import useGetSocialProfile from 'hooks/useGetSocialProfile';
-import useUserProfile from 'hooks/useUserProfile';
-import { mockSocialProfile } from 'constants/mockSocialProfile';
 import type { Colors } from 'styles/theme';
 import type { SocialProfile } from 'typings/social';
 import SocialFollowButton from 'components/social/profile/SocialFollowButton';
@@ -11,33 +7,12 @@ import EditProfileButton from 'components/social/profile/EditProfileButton';
 import SocialProfileImage from 'components/social/profile/SocialProfileImage';
 // import SocialUnFollowButton from 'components/social/profile/SocialUnFollowButton';
 
-export const SocialProfileCard = () => {
-  const router = useRouter();
-  const { pathname } = router;
-  const isMyPage = pathname === '/social/profile';
-
-  const userData = useUserProfile();
-  const { data: socialProfile, status } = useGetSocialProfile(
-    userData?.nickname
-  );
-
-  if (status && socialProfile?.userId) {
-    return (
-      <SocialProfileCardView isMyPage={isMyPage} profile={socialProfile} />
-    );
-  }
-
-  return (
-    <SocialProfileCardView isMyPage={isMyPage} profile={mockSocialProfile} />
-  );
-};
-
 interface SocialProfileCardViewProps {
   isMyPage: boolean;
   profile: SocialProfile;
 }
 
-export const SocialProfileCardView = ({
+const SocialProfileCardView = ({
   isMyPage,
   profile,
 }: SocialProfileCardViewProps) => {
@@ -98,6 +73,8 @@ export const SocialProfileCardView = ({
     </S.ProfileCard>
   );
 };
+
+export default SocialProfileCardView;
 
 const S = {
   ProfileCard: styled.div`
